pageId를 LongWritable로 변경하였음. 테스트 필요.

+++ src/main/java/InvertedIndexDriver.java
@@ -1,7 +1,7 @@
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -33,7 +33,7 @@ public class InvertedIndexDriver extends Configured implements Tool {
 		job.setReducerClass(InvertedIndexReducer.class);
 
 		job.setMapOutputKeyClass(Text.class);
-		job.setMapOutputValueClass(IntWritable.class);
+		job.setMapOutputValueClass(LongWritable.class);
 
 		job.setOutputKeyClass(Text.class);
 		job.setOutputValueClass(CustomArrayWritable.class);
